Рефакторинг кода в form.js
