+++ js/main.js
@@ -1,8 +1,8 @@
 import { getRandomCards } from './data.js';
 import { renderCard } from './services/render.js';
-import { setFormState } from './services/form-state.js';
+import { activateForm, deactivateForm } from './services/form.js';
 const cards = getRandomCards();
 const card = cards[0];
 renderCard(card);
-setFormState(true);
-setTimeout(setFormState, 3000, false);
+deactivateForm(true);
+setTimeout(activateForm, 3000, false);
